datadog/src/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
ddtrace==2.3.0
datadog==0.47.0 